Create Print the Date and Time
"position: absolute" is a CSS property that is used to position an element relative to its nearest positioned ancestor, instead of its parent element. If there is no positioned ancestor, the element will be positioned relative to the initial containing block, which is usually the body element.

The "position: absolute" value takes the element out of the normal document flow, and it will not affect the layout of other elements on the page. Instead, the element can be positioned using the "top", "bottom", "left", and "right" properties. These properties determine the distance of the element from the edges of its nearest positioned ancestor or the initial containing block.

Here is an example of how "position: absolute" can be used:

.container {
  position: relative;
}

.element {
  position: absolute;
  top: 20px;
  right: 20px;
}


In this example, the parent element has a class of "container" and has "position: relative" set. The child element has a class of "element" and has "position: absolute" set. The "top" and "right" properties position the "element" 20 pixels from the top and right edges of the "container" element, respectively.

It's important to note that the "position: absolute" property can cause elements to overlap if they are not positioned correctly, so it should be used with caution